refactor(gpu): share compact stride validation
